chore: Reorganize / rename assets
