+++ .luarc.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "https://raw.githubusercontent.com/sumneko/vscode-lua/master/setting/schema.json",
+  "diagnostics": {
+    "globals": ["vim"],
+    "disable": ["different-requires"]
+  },
+  "runtime": {
+    "version": "LuaJIT"
+  },
+  "workspace": {
+    "maxPreload": 100000,
+    "preloadFileSize": 10000,
+    "ignoreDir": [".vscode", ".git"],
+    "library": ["."]
+  },
+  "telemetry": {
+    "enable": false
+  }
+}
