+++ lua/core/event.lua
@@ -24,7 +24,7 @@ vim.api.nvim_create_autocmd("BufEnter", {
 			and vim.api.nvim_buf_get_option(vim.api.nvim_win_get_buf(layout[2]), "filetype") == "NvimTree"
 			and layout[3] == nil
 		then
-			vim.cmd("confirm quit")
+			vim.api.nvim_command([[confirm quit]])
 		end
 	end,
 })
