+++ lua/modules/configs/completion/servers/sumneko_lua.lua
@@ -8,8 +8,6 @@ return {
 					[vim.fn.expand("$VIMRUNTIME/lua")] = true,
 					[vim.fn.expand("$VIMRUNTIME/lua/vim/lsp")] = true,
 				},
-				maxPreload = 100000,
-				preloadFileSize = 10000,
 			},
 			telemetry = { enable = false },
 			-- Do not override treesitter lua highlighting with sumneko lua highlighting
