+++ lua/modules/configs/completion/tabnine.lua
@@ -1,0 +1,3 @@
+return function()
+	require("cmp_tabnine.config"):setup({ max_line = 1000, max_num_results = 20, sort = true })
+end
