+++ lua/modules/configs/editor/colorizer.lua
@@ -1,0 +1,3 @@
+return function()
+	require("colorizer").setup({})
+end
