--- lua/modules/configs/editor/nvim-colorizer.lua
@@ -1,3 +1,0 @@
-return function()
-	require("colorizer").setup()
-end
