+++ lua/modules/plugins/lang.lua
@@ -4,7 +4,7 @@ lang["fatih/vim-go"] = {
 	lazy = true,
 	ft = "go",
 	build = ":GoInstallBinaries",
-	config = require("lang.golang"),
+	config = require("lang.vim-go"),
 }
 lang["simrat39/rust-tools.nvim"] = {
 	lazy = true,
